6-size.c

+++ 0x00-hello_world/6-size.c
@@ -1,4 +1,4 @@
-#include <stdio.h>
+#include<stdio.h>
 
 /**
  * main - entry point
@@ -13,10 +13,10 @@ int main(void)
 	long int c;
 	long long int d;
 	float f;
-	printf("size of a char: %lu byte(s)\n", sizeof(a));
-	printf("size of a int: %lu byte(s)\n", sizeof(b));
-	printf("size of a long int: %lu byte(s)\n", sizeof(c));
-	printf("size of a long long int: %lu byte(s)\n", sizeof(d));
-	printf("size of a float: %lu byte(s)\n", sizeof(f));
+	printf("size of a char: %lu byte(s)\n",(unsigned long)sizeof(a));
+	printf("size of a int: %lu byte(s)\n", (unsigned long)sizeof(b));
+	printf("size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
+	printf("size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
 	return (0);
 }
